Color change re-implemented
Player can now see the cube in the QR Code and change the color to black when pressing on it.

TO DO:
Test out if these change are also visible on another device that is in the same session. If this is not the case implement a way to display this